Minor fixes.

+++ src/pocklington.cpp
@@ -155,7 +155,7 @@ void Pocklington::run(InputNum& input, arithmetic::GWState& gwstate, File& file_
             }
 
             PrimeIterator primes = PrimeIterator::get();
-            for (; (uint32_t)*primes <= _a; primes++);
+            for (; *primes <= _a; primes++);
             _a = *primes;
             std::string sa = std::to_string(_a);
             if (dynamic_cast<FastExp*>(_task.get()) != nullptr)
